chore: 주석 및 기타 작업
- type 파일 생성
- 주석 작성
- 오타 수정 및 console.log 삭제
