+++ src/apis/giftOrderApi.ts
@@ -1,10 +1,12 @@
 import axiosInstance from './axiosInstance';
 
+//선물 주문 페이지(GiftOrderPage) 상품 요약 정보 요청 API
 export const fetchProductSummury = async (id: string | undefined) => {
   const res = await axiosInstance.get(`/products/${id}/summary`);
   return res.data.data;
 };
 
+// 선물 주문 페이지 주문 요청 API
 interface OrderType {
   body: Record<string, unknown>;
   options: { headers: Record<string, string> };
