+++ src/apis/rankingApi.ts
@@ -1,5 +1,6 @@
 import axiosInstance from './axiosInstance';
 
+// Home 페이지 랭킹 상품 요청 API
 export const fetchRankedProducts = async (
   apiTargetType: string,
   apiRankType: string
