+++ src/apis/themeApi.ts
@@ -1,16 +1,19 @@
+import type { ThemeProducts } from '@features/ThemeProductList/themeProductType';
 import axiosInstance from './axiosInstance';
-import type { ThemeProducts } from '@features/ThemeProductList/components/ProductList';
 
+// Home 페이지 Theme 요청 API
 export const fetchThemes = async () => {
   const res = await axiosInstance.get('/themes');
   return res.data.data;
 };
 
+// Theme Product List 페이지 상단 Hero 정보 요청 API
 export const fetchThemeInfo = async (id: string | undefined) => {
   const res = await axiosInstance.get(`/themes/${id}/info`);
   return res.data.data;
 };
 
+// Theme Product List 페이지 상품 요청 API
 export const fetchThemeProducts = async (
   id: number,
   cursor: number,
