+++ src/features/ThemeProductList/components/ProductList.tsx
@@ -22,20 +22,25 @@ const ProductList = ({ id }: ThemeProductProps) => {
   } = useInfiniteQuery<ThemeProducts, Error>({
     queryKey: ['themeProducts', id],
 
+    // 데이터 요청함수
     queryFn: ({ pageParam = 0 }: QueryFunctionContext) =>
       fetchThemeProducts(Number(id), pageParam as number),
 
+    // 다음 페이지를 가져올 기준
     getNextPageParam: (lastPage) =>
       lastPage.hasMoreList ? lastPage.cursor : undefined,
+
+    // 초기 페이지 Param
     initialPageParam: 0,
   });
 
+  // 무한 스크롤 훅 사용
   const { sentinelRef } = useInfiniteScroll({
     onIntersect: fetchNextPage,
     enabled: !!hasNextPage && !isFetchingNextPage,
   });
 
-  console.log(data);
+  // 모든 페이지에서 상품 리스트를 평탄화
   const products = data?.pages.flatMap((page) => page.list) ?? [];
 
   if (isPending) return <LoadingSpinner />;
@@ -52,11 +57,11 @@ const ProductList = ({ id }: ThemeProductProps) => {
 
   return (
     <>
-      <GridWrqpper>
+      <GridWrapper>
         {products.map((item, index) => (
           <ProductCard key={index} {...item} />
         ))}
-      </GridWrqpper>
+      </GridWrapper>
       {hasNextPage && <div ref={sentinelRef} style={{ height: '1px' }} />}
     </>
   );
@@ -64,7 +69,7 @@ const ProductList = ({ id }: ThemeProductProps) => {
 
 export default ProductList;
 
-const GridWrqpper = styled.div(({ theme }) => ({
+const GridWrapper = styled.div(({ theme }) => ({
   display: 'grid',
   gridTemplateColumns: 'repeat(3, 1fr)',
   gap: theme.spacing.spacing5,
