+++ src/types/auth.ts
@@ -1,0 +1,10 @@
+export interface LoginCredentials {
+  email: string;
+  password: string;
+}
+
+export interface User {
+  email: string;
+  name: string;
+  authToken: string;
+}
